ecommerce_backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.0
django-cors-headers==4.7.0 